Check each incoming SMS for a STOP opt-out request
In the incoming SMS handler, check each message to see if it contains
only the case-insensitive word "STOP".

If it does, handle the opt out by removing the sender from the chat and
deleting the chat connection.

The sender of the opt-out request will be notified that they will
no longer receive messages, and the remaining chat room members will
get a notice announcing who left the chat.